Configure the server for testing
