+++ server/__tests__/api.test.js
@@ -1,0 +1,6 @@
+const request = require('supertest');
+const app = require('../app');
+
+it('should be false', async () => {
+  return 1 === 1;
+});
